Factor out location cache and counters to their own file
